% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_methods.R
\name{euclidean.dist}
\alias{euclidean.dist}
\title{Euclidean Dissimilarity}
\usage{
euclidean.dist(x, y)
}
\description{
The canonical dissimilarity function, the euclidean distance metric, also known as l2 distance.
}
