% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_methods.R
\name{inner.prod.dist}
\alias{inner.prod.dist}
\title{Inner product Dissimilarity}
\usage{
inner.prod.dist(x, y)
}
\description{
A dissimilarity function that is simply the inner product of the two inputs. Note that this is not a metric.
}
